requirements.txt
absl-py==0.13.0
attrs==21.2.0
certifi==2021.5.30
charset-normalizer==2.0.4
clearml==1.0.5
deepdiff==5.5.0
enum34==1.1.10
furl==2.1.2
future==0.18.2
humanfriendly==9.2
idna==3.2
jsonschema==3.2.0
mock==4.0.3
mpyq==0.2.5
numpy==1.21.2
ordered-set==4.0.2
orderedmultidict==1.0.1
pathlib2==2.3.6
Pillow==8.3.2
portpicker==1.4.0
protobuf==3.17.3
psutil==5.8.0
pygame==2.0.1
PyJWT==2.1.0
pyparsing==2.4.7
pyrsistent==0.18.0
PySC2==3.0.0
python-dateutil==2.8.2
PyYAML==5.4.1
requests==2.26.0
s2clientprotocol==5.0.7.84643.0
s2protocol==5.0.7.84643.0
sc2reader==1.7.0
scipy==1.7.1
six==1.16.0
sk-video==1.1.10
torch==1.9.0+cu111
typing-extensions==3.10.0.2
urllib3==1.26.6
websocket-client==1.2.1
whichcraft==0.6.1
zstd==1.5.0.2
